authorisation

+++ src/main/java/ch/bzz/kino/service/FilmService.java
@@ -9,6 +9,7 @@ import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.Pattern;
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.NewCookie;
 import javax.ws.rs.core.Response;
 import java.util.List;
 import java.util.UUID;
@@ -31,19 +32,30 @@ public class FilmService {
     public Response listFilme(
             @CookieParam("userRole") String userRole
     ) {
-        List<Film> filmList = DataHandler.getInstance().readAllFilme();{
-            int httpStatus;
-            if (userRole == null || userRole.equals("guest")) {
-                httpStatus = 404;
-            } else {
-                httpStatus = 200;
-            }
-            return Response
-                    .status(httpStatus)
-                    .status(200)
-                    .entity(filmList)
-                    .build();
+        List<Film> filmList = null;
+        int httpStatus;
+        if (userRole == null || userRole.equals("guest")) {
+            httpStatus = 404;
+        } else {
+            httpStatus = 200;
+            filmList = DataHandler.getInstance().readAllFilme();
         }
+
+        NewCookie cookie = new NewCookie(
+                "userRole",
+                userRole,
+                "/",
+                "",
+                "Login-Cookie",
+                600,
+                false
+        );
+
+        return Response
+                .status(httpStatus)
+                .entity(filmList)
+                .cookie(cookie)
+                .build();
     }
 
     /**
@@ -56,6 +68,7 @@ public class FilmService {
     @Produces(MediaType.APPLICATION_JSON)
     public Response readFilm(
             @NotEmpty
+            @CookieParam("userRole") String userRole,
             @Pattern(regexp = "[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}")
             @PathParam("uuid") String filmUUID
 
@@ -68,9 +81,20 @@ public class FilmService {
                     .build();
         }
 
+        NewCookie cookie = new NewCookie(
+                "userRole",
+                userRole,
+                "/",
+                "",
+                "Login-Cookie",
+                600,
+                false
+        );
+
         return Response
                 .status(200)
                 .entity(film)
+                .cookie(cookie)
                 .build();
     }
 
@@ -84,16 +108,29 @@ public class FilmService {
     @Path("create")
     @Produces (MediaType.TEXT_PLAIN)
     public Response insertFilm(
+            @NotEmpty
             @Valid @BeanParam Film film,
+            @CookieParam("userRole") String userRole,
             @FormParam("filmUUID") String filmUUID
     ){
         film.setFilmUUID(filmUUID);
         film.setFilmUUID(UUID.randomUUID().toString());
-
         DataHandler.getInstance().insertFilm(film);
+
+        NewCookie cookie = new NewCookie(
+                "userRole",
+                userRole,
+                "/",
+                "",
+                "Login-Cookie",
+                600,
+                false
+        );
+
         return Response
                 .status(200)
-                .entity("Film succesfully iserted")
+                .entity("Film succesfully inserted")
+                .cookie(cookie)
                 .build();
     }
 
@@ -107,11 +144,13 @@ public class FilmService {
     @Path("update")
     @Produces (MediaType.TEXT_PLAIN)
     public Response updateFilm(
+            @NotEmpty
             @Valid @BeanParam Film film,
-            @FormParam("filmUUID") String filmUUID
-
+            @FormParam("filmUUID") String filmUUID,
+            @CookieParam("userRole") String userRole
     ){
         int httpStatus = 200;
+
         Film oldFilm = DataHandler.getInstance().readFilmByUUID(film.getFilmUUID());
         if (film == null) {
             oldFilm.setTitel(film.getTitel());
@@ -124,9 +163,21 @@ public class FilmService {
         }else {
             httpStatus = 404;
         }
+
+        NewCookie cookie = new NewCookie(
+                "userRole",
+                userRole,
+                "/",
+                "",
+                "Login-Cookie",
+                600,
+                false
+        );
+
         return Response
                 .status(httpStatus)
                 .entity("Film succesfully updated")
+                .cookie(cookie)
                 .build();
     }
 
@@ -141,15 +192,31 @@ public class FilmService {
     public Response deleteFilm(
             @NotEmpty
             @Pattern(regexp = "[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}")
-            @QueryParam("uuid") String filmUUID
+            @FormParam("uuid") String filmUUID,
+            @CookieParam("userRole") String userRole
     ) {
         int httpStatus = 200;
+        if(userRole.equals("admin"))
         if (!DataHandler.getInstance().deleteFilm(filmUUID)) {
             httpStatus = 404;
+        } else {
+            httpStatus = 200;
         }
+
+        NewCookie cookie = new NewCookie(
+                "userRole",
+                userRole,
+                "/",
+                "",
+                "Login-Cookie",
+                600,
+                false
+        );
+
         return Response
                 .status(httpStatus)
                 .entity("film deleted")
+                .cookie(cookie)
                 .build();
     }
